Added notification-service and made other changes

+++ order-service/src/main/java/com/example/os/api/repository/OrderRepository.java
@@ -3,6 +3,5 @@ package com.example.os.api.repository;
 import com.example.os.api.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
-
+public interface OrderRepository extends JpaRepository<Order, Long> {
 }
